packages/canvas-workflow-kit/canvas_workflow_kit-0.6.11-py3-none-any.whl/canvas_workflow_kit/builtin_cqms/helper_population.py
class Population:

    def __init__(self):
        self.in_initial_population: bool = None
        self.in_denominator: bool = None
        self.in_numerator: bool = None

    def set_initial_population(self, flag: bool):
        self.in_initial_population = flag
        self.in_denominator = flag
        self.in_numerator = flag

    def set_denominator(self, flag: bool):
        self.in_denominator = flag
        self.in_numerator = flag

    def set_numerator(self, flag: bool):
        self.in_numerator = flag
